Revert "DO NOT MERGE Disable overlays while installer is resumed"
If mOkCanInstall is false when onResume() is called, the button gets
disabled and never re-enabled. This fix is for branches that have
change I42c15d05dffc1f6db69c9f67148f7ceb6e8bd8ae ("Split layout for
confirm step.") which causes the button to be disabled and enabled
for different reasons. Here the button is constantly enabled and its
action changes depending on the value of mOkCanInstall. The attribute
filterTouchesWhenObscured should prevent tapjacking when overlays are
present, so revert this commit.

This has also been done in branch cm-14.1 via commit
4b0c302eae8d772d09346718559cc42548485f75

This reverts commit 7a204a0e265a0a92537d57fd1bab594250c96149.

Change-Id: I82b2c9ae0d588ef1696c8356a31c4204247823ad